First upload linux dev version

+++ libs/boost/thread/pthread/thread.cpp
@@ -377,7 +377,7 @@ namespace boost
         return pthread_num_processors_np();
 #elif defined(__APPLE__) || defined(__FreeBSD__)
         int count;
-        size_t size=sizeof(count);
+        unsigned int size=sizeof(count);
         return sysctlbyname("hw.ncpu",&count,&size,NULL,0)?0:count;
 #elif defined(BOOST_HAS_UNISTD_H) && defined(_SC_NPROCESSORS_ONLN)
         int const count=sysconf(_SC_NPROCESSORS_ONLN);
